requirements.txt
azure-cli
azure-identity>=1.10.0
azure-mgmt-resource>=22.0.0
azure-mgmt-subscription>=3.1.1
azure-mgmt-compute>=27.0.0
azure-mgmt-network>=22.0.0
azure-mgmt-storage>=20.0.0
azure-mgmt-keyvault>=10.0.0
azure-mgmt-monitor>=5.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0
black>=22.3.0
pylint>=2.13.0
flake8>=4.0.1
colorama==0.4.6
pyfiglet==0.8.post1
tqdm==4.66.1